refactor: extract message-input logic into reusable mixin
