+++ packages/message-input/package.json
@@ -35,6 +35,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/button": "24.6.0-beta1",
     "@vaadin/component-base": "24.6.0-beta1",
